server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
django==5.1.7
django-cors-headers==4.3.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-types==0.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
djangorestframework-stubs==3.15.3
idna==3.10
pillow==11.1.0
pyjwt==2.8.0
requests==2.32.3
sqlparse==0.5.3
types-psycopg2==2.9.21.20250318
types-pyyaml==6.0.12.20250402
types-requests==2.32.0.20250328
typing-extensions==4.13.2
urllib3==2.4.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
python-dotenv==1.0.1
